Implement search with ember-bootstrap components
Previous implementation used the native bootstrap HTML elements.
These elements don't have the necessary JavaScript logic attached to
them.  As ember-bootstrap maintains this fuctionality, we choose not
to replicate, and to implement the search based on their components.

The searchPlaceholder computed property could be used as the
'placeholder' property on the search input.  It is not implemented at
this stage.  Kees may decide what works best.

We removed the bootstrap scss import in the styling for now, as it
doesn't seem to impact the rendered design.