First stab at using ember concurrency for page fetching
The Ember router is great for fetching standard contents.  When
fetching contents from a different type of API, or when we need more
control over the rendered contents (as in: our UI doesn't match the
traditional way of using routes), Ember Concurrency provides a nice
way to opt out of the standard way of working.  This commit provides
the first shift towards using Ember Concurrency.