Search automatically when opening page with arguments
The search now automatically launches when arguments were given on
page load.  This allows users to share direct links to searches.  The
implementation of this was a good place to ensure searches for "any"
type without search string would not yield results any longer.  It is
still possible to search for any music, video, directory, or file.

Without specifying the type of search it is now required to enter a
search string, as the API requires.

resolves ipfs-search/ipfs-search-frontend#3