Disabling package-lock.json file
In #20 the package-lock.json file is discussed.  The conclusion is
that reproducible builds will make it easiers for contributors to join
in.  Given that the lockfile is different in semi-common NPM versions,
and that we don't have a lock on the NPM version itself yet, the
package-lock.json file is disabled.