Implemented a clickable-menu-items css class for dropdown-menu
It is advised to have menu items in the bootstrap dropdown-menu
contain an anchor tag.  Alhough that would be prettier, we don't use
that, we maintain state in a less clean way at the moment.  Triggering
of actions is done through click events.  IOs does not fire these
events unless there's a `cursor:pointer` style on the relevant
element.  THe clickable-menu-items thus adds this cursor:pointer to
the respective element.