Making dropdown button white
I'm not sure about this one.  The CSS seems to give me the impression
that the button should have been transparent, yet the documentation
states the inverse.  Trying out some things, I believe it was the goal
to have the dropdown button in a traditional white background.

The previous implementation of this did not seem to work correctly on
IOs.  Now we're consistent.  If the implementation should have been
different, we'll also manage that.