Get rid of changes made in the last 5 commits
The last five commits were an attempt to solve issue #38 (https://github.com/ipfs-search/ipfs-search-frontend/pull/38). We needed to highlight the last previewed item. However, the implementation had tried to go furthur. It would highlight older previewed items in a different color as well. The problem with this implementation is that it basedd on sharing component instance references. This is seen in the openModal action of the searchhit-card component. Here, this (a reference to the current component) is sent to the calling component (in this case, search-results). THIS IS DANGEROUSLY WRONG.