Highlight last previewed item (#38)
The last item that was previewed is now marked, making it easier to track.

* Create necessary classes

The 'selected' class will be assigned to the last previewd and the 'selected-old' calss will be assigned to other previewd results.

* Update classList when item is previewed

* Keep track of what has been previewed

* Logic for updating the classes

* Upgrade ember-bootstrap addon's version

* Get rid of changes made in the last 5 commits

The last five commits were an attempt to solve issue #38 (https://github.com/ipfs-search/ipfs-search-frontend/pull/38). We needed to highlight the last previewed item. However, the implementation had tried to go furthur. It would highlight older previewed items in a different color as well. The problem with this implementation is that it basedd on sharing component instance references. This is seen in the openModal action of the searchhit-card component. Here, this (a reference to the current component) is sent to the calling component (in this case, search-results). THIS IS DANGEROUSLY WRONG.

* Add  property to search-result

This property will keep track of the last selected/visited search result. It will be set to that result's hash

* Add  action to search-results

This will update the current selected result record.

* Inject  action to search-hit-card

This will update the current selected result record when the card is clicked.

* Add a new `selected` propety and bind style to it

* Bind  to

This will update the  value (from search-hit-card) to the  value (from searchresult) which contains the hash of last visited result. This will in turn update the style of search-hit-card.

* Replace the redundant setAsSelectedHit with a mut

* Rename onclick to onOpenDetail

* Check onOpenDetails validity before calling it

onOpenDetails is a callback argument. So, before calling it, we need to make sure it is valid.

* Change hover color & last selected item's bg color