Use router service instead of rolling our own
We had the intention of doing fancy animations and all that.  That
lead us to implementing our own Router logic so that would be a more
feasible thing to add.  All of that did not happen so we're using the
router directly instead.

There is more cleanup possible than what we did here.  Given the
current code, we may want to switch to a loading and error template
instead as we don't have animations in these cases at the moment
either.