Support multiple video gateways
As suggested in #45, offering multiple video gateways will allow us to
embed the videos.  We now support both the mainstream gateway.ipfs.io,
as well as cloudflare's ipfs gateway.  The combination of these two
should provide better support.

We could use a similar approach in other places.  This is the one
where the effect is the most apparent.